Remove Args from MustDestroy type parameters
Unfortunately, knowing the args as part of the MustDestroy struct makes
usage a bit of a pain in the arse.

However, having now removed it, we have to pass args to destroy method
as a tuple. The way I tried to hide it previously, doesn't appear to work
and will likely require some clever type level tricks than I am
currently capable of reasoning about.